client/src/app/components/table/delete-button/delete-button.component.spec.ts
import { HttpClientModule } from '@angular/common/http';
import { ComponentFixture, TestBed } from '@angular/core/testing';

import { DeleteButtonComponent } from './delete-button.component';

describe('DeleteButtonComponent', () => {
  let component: DeleteButtonComponent;
  let fixture: ComponentFixture<DeleteButtonComponent>;
  let spy: any;

  beforeEach(async () => {
    await TestBed.configureTestingModule({
      imports: [HttpClientModule],
      declarations: [ DeleteButtonComponent ]
    })
    .compileComponents();
  });

  beforeEach(() => {
    fixture = TestBed.createComponent(DeleteButtonComponent);
    component = fixture.componentInstance;
    spy = spyOn(component, "deleteHandler");
    fixture.detectChanges();
  });

  it('should create', () => {
    expect(component).toBeTruthy();
  });

  it("deletes game", async () => {
    const button = fixture.nativeElement.querySelector("button");
    expect(spy).not.toHaveBeenCalled();
    button.click();
    fixture.detectChanges();
    await new Promise(r => setTimeout(r, 50));
    expect(spy).toHaveBeenCalledTimes(1);
  });
});
